split sequences into own router, pulled into data router
